pyproject.toml
[project]
name = "jobs"
version = "0.2.0"
description = "US Job Market Visualizer - BLS OOH data pipeline with AI exposure scoring"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tenacity>=8.2.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
jobs = "jobs.cli:main"
